more tests + housekeeping (#62)
* more tests + housekeeping

* truthy = integrations

* SKIP_INTEGRATION_TESTS
