--- .bandit.yml
@@ -1,5 +1,0 @@
----
-skips: []
-# No need to check for security issues in the test scripts!
-exclude_dirs:
-  - "./tests/"
