+++ tests/test_data/schema_path/nokia-conf-aaa.yang
@@ -1,1 +1,1 @@
-<MagicMock name='connect_ssh().get_schema()' id='140064057824912'>
+<MagicMock name='connect_ssh().get_schema()' id='140033250936912'>
